+++ mod.ts
@@ -1,1 +1,2 @@
 export * from "./src/pup.ts";
+export * from "./src/rest.ts";