+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.5] - 2024-05-02
+
+### Added
+
+- Add and export Error type `RestAPIConnectionError`
+- Add timeout
+
+### Changed
+
+- Export Error type `RestAPIError`
+
 ## [1.0.4] - 2024-04-30
 
 ### Added